requirements.txt
pyTenable>=1.4.8
pytest>=3.6
